[All]: 修改函数名、调整部分代码

+++ src/exec/crawler/incl/crawler.h
@@ -22,18 +22,18 @@
 #include "thread_pool.h"
 
 /* 宏定义 */
-#define CRWL_TMOUT_SEC              (02)        /* 超时(秒) */
-#define CRWL_TMOUT_USEC             (00)        /* 超时(微妙) */
-#define CRWL_DEF_THD_NUM            (01)        /* 默认线程数 */
-#define CRWL_SLAB_SIZE              (1 * MB)    /* SLAB内存池大小 */
+#define CRWL_TMOUT_SEC              (02)    /* 超时(秒) */
+#define CRWL_TMOUT_USEC             (00)    /* 超时(微妙) */
+#define CRWL_DEF_THD_NUM            (01)    /* 默认线程数 */
+#define CRWL_SLAB_SIZE              (1 * MB)/* SLAB内存池大小 */
 #define CRWL_RECV_SIZE              (128 * KB)  /* 缓存SIZE(接收缓存) */
 #define CRWL_SYNC_SIZE              (64 * KB)   /* 同步SIZE */
-#define CRWL_DOWN_WEBPAGE_NUM       (1)         /* 默认同时下载的网页数 */
-#define CRWL_CONNECT_TMOUT_SEC      (00)        /* 连接超时时间 */
-#define CRWL_WEB_SVR_PORT           (80)        /* WEB服务器侦听端口 */
-#define CRWL_SCK_TMOUT_SEC          (05)        /* 套接字超时时间(秒) */
+#define CRWL_DOWN_WEBPAGE_NUM       (1)     /* 默认同时下载的网页数 */
+#define CRWL_CONNECT_TMOUT_SEC      (00)    /* 连接超时时间 */
+#define CRWL_WEB_SVR_PORT           (80)    /* WEB服务器侦听端口 */
+#define CRWL_SCK_TMOUT_SEC          (05)    /* 套接字超时时间(秒) */
 
-#define CRWL_TASK_QUEUE_MAX_NUM     (10000)     /* 任务队列单元数 */
+#define CRWL_TASK_QUEUE_MAX_NUM     (10000) /* 任务队列单元数 */
 #define CRWL_DEF_CONF_PATH  "../conf/crawler.xml"   /* 默认配置路径 */
 
 
@@ -41,7 +41,7 @@
 /* 错误码 */
 typedef enum
 {
-    CRWL_OK = 0
+    CRWL_OK = 0                             /* 正常 */
     , CRWL_SHOW_HELP                        /* 显示帮助信息 */
 
     , CRWL_ERR = ~0x7fffffff                /* 失败、错误 */
@@ -110,11 +110,12 @@ typedef struct
     eslab_pool_t slab;                      /* 内存池 */
 } crwl_cntx_t;
 
+/* 对外接口 */
 crwl_cntx_t *crwl_cntx_init(const crwl_conf_t *conf, log_cycle_t *log);
 int crwl_cntx_startup(crwl_cntx_t *ctx);
 
 int crwl_slab_init(crwl_cntx_t *ctx);
-void *crwl_slab_alloc(crwl_cntx_t *ctx, int size);
+void *crwl_slab_alloc(crwl_cntx_t *ctx, size_t size);
 void crwl_slab_dealloc(crwl_cntx_t *ctx, void *p);
 void crwl_slab_destroy(crwl_cntx_t *ctx);
